functions_portfolio/requirements.txt
firebase_functions~=0.5.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.10.0
google-cloud-pubsub>=2.18.0
google-genai>=1.0.0
yfinance>=0.2.28
numpy>=1.24.0
python-dotenv>=1.0.0
